labs/py/requirements.txt
azure-ai-projects==2.0.0b3
azure-identity>=1.15.0
openai>=2.8.0
python-dotenv>=1.0.0
azure-mgmt-cognitiveservices>=14.0.0
azure-monitor-opentelemetry>=1.6.0
opentelemetry-api>=1.20.0
